+++ .github/workflows/ci.yml
@@ -55,10 +55,15 @@ jobs:
     - run: cargo clippy --version
     - run: cargo clippy
     - run: cargo clippy --all-targets --no-default-features
-    - run: cargo clippy --all-targets --features native-tls
-    - run: cargo clippy --all-targets --features rustls-tls
-    - run: cargo clippy --all-targets --features rustls-tls-aws
     - run: cargo clippy --all-targets --all-features
+
+    # TLS
+    - run: cargo clippy --features native-tls
+    - run: cargo clippy --features rustls-tls
+    - run: cargo clippy --features rustls-tls-ring,rustls-tls-webpki-roots
+    - run: cargo clippy --features rustls-tls-ring,rustls-tls-native-roots
+    - run: cargo clippy --features rustls-tls-aws-lc,rustls-tls-webpki-roots
+    - run: cargo clippy --features rustls-tls-aws-lc,rustls-tls-native-roots
 
   test:
     runs-on: ubuntu-latest
